Redo to ts datepicker components
