--- lib/src/DatePicker/components/Day.d.ts
@@ -1,14 +1,0 @@
-import { ComponentClass, ReactNode } from 'react';
-
-export interface DayProps {
-    children: ReactNode;
-    current?: boolean;
-    disabled?: boolean;
-    hidden?: boolean;
-    selected?: boolean;
-}
-
-declare const Day: ComponentClass<DayProps>;
-
-export default Day;
-
