+++ lib/src/DatePicker/components/SlideTransition.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import * as React from 'react';
+import * as PropTypes from 'prop-types';
 import classnames from 'classnames';
 import withStyles from '@material-ui/core/styles/withStyles';
 import TransitionGroup from 'react-transition-group/TransitionGroup';
