+++ lib/src/DateTimePicker/DateTimePickerInline.tsx
@@ -1,37 +1,36 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import * as React from 'react';
+import * as PropTypes from 'prop-types';
 
-import InlineWrapper from '../wrappers/InlineWrapper';
-import DatePicker from './DatePicker';
 import DomainPropTypes from '../constants/prop-types';
+import InlineWrapper from '../wrappers/InlineWrapper';
+import DateTimePicker from './DateTimePicker';
 import BasePicker from '../_shared/BasePicker';
-import Calendar from './components/Calendar';
 
-export const DatePickerInline = (props) => {
+export const DateTimePickerInline = (props) => {
   const {
-    allowKeyboardControl,
-    animateYearScrolling,
-    disableFuture,
-    disablePast,
-    format,
-    forwardedRef,
-    labelFunc,
-    leftArrowIcon,
-    maxDate,
-    minDate,
-    initialFocusedDate,
-    onChange,
-    openToYearSelection,
-    renderDay,
-    rightArrowIcon,
-    shouldDisableDate,
     value,
+    format,
     autoOk,
-    onlyCalendar,
+    openTo,
+    minDate,
+    maxDate,
+    initialFocusedDate,
+    showTabs,
+    autoSubmit,
+    disablePast,
+    disableFuture,
+    leftArrowIcon,
+    rightArrowIcon,
+    dateRangeIcon,
+    timeIcon,
+    renderDay,
+    ampm,
+    shouldDisableDate,
+    animateYearScrolling,
+    forwardedRef,
+    allowKeyboardControl,
     ...other
   } = props;
-
-  const ComponentToShow = onlyCalendar ? Calendar : DatePicker;
 
   return (
     <BasePicker {...props} autoOk>
@@ -39,39 +38,44 @@ export const DatePickerInline = (props) => {
         ({
           date,
           utils,
-          isAccepted,
           handleChange,
           handleTextFieldChange,
+          isAccepted,
+          pick12hOr24hFormat,
           handleAccept,
         }) => (
           <InlineWrapper
+            innerRef={forwardedRef}
             disableFuture={disableFuture}
             disablePast={disablePast}
-            format={format || utils.dateFormat}
-            labelFunc={labelFunc}
             maxDate={maxDate}
             minDate={minDate}
             onChange={handleTextFieldChange}
-            innerRef={forwardedRef}
             value={value}
             isAccepted={isAccepted}
             handleAccept={handleAccept}
+            format={pick12hOr24hFormat(utils.dateTime12hFormat, utils.dateTime24hFormat)}
             {...other}
           >
-            <ComponentToShow
-              date={date}
+            <DateTimePicker
               allowKeyboardControl={allowKeyboardControl}
+              ampm={ampm}
               animateYearScrolling={animateYearScrolling}
+              autoSubmit={autoSubmit}
+              date={date}
+              dateRangeIcon={dateRangeIcon}
               disableFuture={disableFuture}
               disablePast={disablePast}
               leftArrowIcon={leftArrowIcon}
               maxDate={maxDate}
               minDate={minDate}
               onChange={handleChange}
-              openToYearSelection={openToYearSelection}
+              openTo={openTo}
               renderDay={renderDay}
               rightArrowIcon={rightArrowIcon}
               shouldDisableDate={shouldDisableDate}
+              showTabs={showTabs}
+              timeIcon={timeIcon}
             />
           </InlineWrapper>
         )
@@ -80,49 +84,55 @@ export const DatePickerInline = (props) => {
   );
 };
 
-DatePickerInline.propTypes = {
-  onlyCalendar: PropTypes.bool,
+DateTimePickerInline.propTypes = {
   value: DomainPropTypes.date,
+  format: PropTypes.string,
+  onChange: PropTypes.func.isRequired,
+  autoOk: PropTypes.bool,
+  autoSubmit: PropTypes.bool,
+  disableFuture: PropTypes.bool,
+  disablePast: PropTypes.bool,
   minDate: DomainPropTypes.date,
   maxDate: DomainPropTypes.date,
   initialFocusedDate: DomainPropTypes.date,
-  format: PropTypes.string,
-  onChange: PropTypes.func.isRequired,
-  disablePast: PropTypes.bool,
-  disableFuture: PropTypes.bool,
-  animateYearScrolling: PropTypes.bool,
-  openToYearSelection: PropTypes.bool,
-  labelFunc: PropTypes.func,
+  showTabs: PropTypes.bool,
   leftArrowIcon: PropTypes.node,
   rightArrowIcon: PropTypes.node,
+  dateRangeIcon: PropTypes.node,
+  timeIcon: PropTypes.node,
   renderDay: PropTypes.func,
+  ampm: PropTypes.bool,
   shouldDisableDate: PropTypes.func,
+  animateYearScrolling: PropTypes.bool,
+  openTo: PropTypes.oneOf(['year', 'date', 'hour', 'minutes']),
   allowKeyboardControl: PropTypes.bool,
   forwardedRef: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
-  autoOk: PropTypes.bool,
 };
 
-DatePickerInline.defaultProps = {
+DateTimePickerInline.defaultProps = {
   value: new Date(),
   format: undefined,
+  autoOk: false,
+  autoSubmit: true,
+  openTo: 'date',
+  disableFuture: false,
+  disablePast: false,
   minDate: '1900-01-01',
   maxDate: '2100-01-01',
   initialFocusedDate: undefined,
-  disableFuture: false,
-  disablePast: false,
-  animateYearScrolling: false,
-  openToYearSelection: false,
-  allowKeyboardControl: true,
+  showTabs: true,
   leftArrowIcon: 'keyboard_arrow_left',
   rightArrowIcon: 'keyboard_arrow_right',
+  dateRangeIcon: 'date_range',
+  timeIcon: 'access_time',
   renderDay: undefined,
-  labelFunc: undefined,
+  ampm: true,
   shouldDisableDate: undefined,
+  animateYearScrolling: false,
   forwardedRef: undefined,
-  autoOk: undefined,
-  onlyCalendar: false,
+  allowKeyboardControl: true,
 };
 
-export default React.forwardRef((props, ref) => (
-  <DatePickerInline {...props} forwardedRef={ref} />
-));
+export default React.forwardRef(
+  (props, ref) => <DateTimePickerInline {...props} forwardedRef={ref} />,
+);
