+++ lib/src/_shared/MaskedInput.tsx
@@ -1,8 +1,8 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
+import * as React from 'react';
+import * as PropTypes from 'prop-types';
 import MaskedInput from 'react-text-mask';
 
-export default class Input extends PureComponent {
+export default class Input extends React.PureComponent {
   static propTypes = {
     mask: PropTypes.any,
     inputRef: PropTypes.func.isRequired,
