+++ lib/src/_shared/ModalDialog.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import * as React from 'react';
+import * as PropTypes from 'prop-types';
 import classnames from 'classnames';
 import EventListener from 'react-event-listener';
 import withStyles from '@material-ui/core/styles/withStyles';
