--- lib/src/wrappers/InlineWrapper.d.ts
@@ -1,14 +1,0 @@
-import { ComponentClass, ReactNode } from 'react';
-import { DateTextFieldProps } from '../_shared/DateTextField';
-import { PopoverProps } from '@material-ui/core/Popover';
-
-export interface InlineWrapperProps extends Partial<DateTextFieldProps> {
-  onOpen?: () => void;
-  onClose?: () => void;
-  handleAccept: () => void;
-  PopoverProps?: Partial<PopoverProps>;
-}
-
-declare const InlineWrapper: ComponentClass<InlineWrapperProps>;
-
-export default InlineWrapper;
